card-ventas
se cambio la vista de las ventas de usar tablas a tarjetas

+++ empleados/actions.php
@@ -1,5 +1,12 @@
 <?php 
     require_once "../conexion.php";
+
+    var_dump($_POST);
+    var_dump($_GET);
+    die;
+
+    if(isset($_POST['Eliminar'])){
+
         $dato=$_GET['delete'];
    
         $delete= "DELETE FROM ventas where IDventa=$dato";
@@ -11,4 +18,16 @@
             header("location: veVentas.php");
         }
 
+
+    }
+
+    if(isset($_POST['Detalles'])){
+        
+        $info=$_GET['info'];
+   
+        $query= "select * from subventas where idventa = $dato";
+        $result= mysqli_query($conexion,$query);
+
+    }
+ 
 ?>
